+++ lib/src/extensions/duration_ext.dart
@@ -10,7 +10,7 @@ extension ExtendedDuration on Duration {
 
     final int hours = e.inHours;
 
-    final String hoursString = hours.remainder(24).toInt().twoDigitTime;
+    final String hoursString = hours.remainder(24).twoDigitTime;
 
     return hoursString;
   }
@@ -20,7 +20,7 @@ extension ExtendedDuration on Duration {
 
     final int secs = e.inSeconds;
 
-    final String secString = secs.remainder(60).toInt().twoDigitTime;
+    final String secString = secs.remainder(60).twoDigitTime;
 
     return secString;
   }
@@ -30,7 +30,7 @@ extension ExtendedDuration on Duration {
 
     final int mins = e.inMinutes;
 
-    final String minString = mins.remainder(60).toInt().twoDigitTime;
+    final String minString = mins.remainder(60).twoDigitTime;
 
     return minString;
   }
@@ -49,18 +49,18 @@ extension ExtendedDuration on Duration {
   }
 
   int get hours {
-    return inHours.remainder(12).toInt();
+    return inHours.remainder(12);
   }
 
   int get minutes {
-    return inMinutes.remainder(60).toInt();
+    return inMinutes.remainder(60);
   }
 
   int get seconds {
-    return inSeconds.remainder(60).toInt();
+    return inSeconds.remainder(60);
   }
 
   int get milliseconds {
-    return inMilliseconds.remainder(60000).toInt();
+    return inMilliseconds.remainder(60000);
   }
 }
